+++ src/main/java/com/techstud/sch_parser/domain/impl/PgupsParser.java
@@ -4,10 +4,12 @@ import com.techstud.sch_parser.domain.Parser;
 import com.techstud.sch_parser.model.Schedule;
 import com.techstud.sch_parser.model.kafka.request.ParsingTask;
 import com.techstud.sch_parser.service.MappingService;
+import com.techstud.sch_parser.service.MappingServiceRef;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 import java.text.MessageFormat;
@@ -15,10 +17,13 @@ import java.util.List;
 
 @Component("PGUPS")
 @Slf4j
-@RequiredArgsConstructor
 public class PgupsParser implements Parser {
 
-    private final MappingService mappingService;
+    private final MappingServiceRef<List<Document>> mappingService;
+
+    public PgupsParser(@Qualifier("pgupsMappingServiceImpl") MappingServiceRef<List<Document>> mappingService) {
+        this.mappingService = mappingService;
+    }
 
     @Override
     public Schedule parseSchedule(ParsingTask task) throws Exception {
@@ -32,6 +37,6 @@ public class PgupsParser implements Parser {
                 .get();
         Document oddWeekDocument = Jsoup.connect(oddUrl).userAgent(userAgent).referrer(referrer).get();
         log.info("Successfully fetching data from PGSPU API");
-        return mappingService.mapPgupsToSchedule(List.of(evenWeekDocument, oddWeekDocument));
+        return mappingService.map(List.of(evenWeekDocument, oddWeekDocument));
     }
 }
