+++ src/main/java/com/techstud/sch_parser/service/MappingService.java
@@ -12,13 +12,9 @@ import java.util.List;
 public interface MappingService {
     Schedule mapSseuToSchedule(List<SseuApiResponse> weekSseuSchedules);
 
-    Schedule mapSsauToSchedule(List<Document> documents);
-
     Schedule mapBmstuToSchedule(BmstuApiResponse bmstuApiResponse);
 
     Schedule mapTltsuToSchedule(List<TltsuApiResponse> documents) throws EmptyScheduleException;
-
-    Schedule mapPgupsToSchedule(List<Document> documents);
 
     Schedule mapSpbstuScheduleByScheduleDay(List<Document> documents);
 }
