+++ src/main/java/com/techstud/sch_parser/service/impl/MappingServiceImpl.java
@@ -72,31 +72,6 @@ public class MappingServiceImpl implements MappingService {
     }
 
     @Override
-    public Schedule mapSsauToSchedule(List<Document> documents) {
-        log.info("Start mapping SSAU data to schedule");
-        Element evenElement = documents.get(0)
-                .getElementsByClass("schedule")
-                .first();
-
-        Element oddElement = documents.get(1)
-                .getElementsByClass("schedule")
-                .first();
-
-        if (evenElement == null && oddElement == null) {
-            return null;
-        }
-
-        Map<DayOfWeek, ScheduleDay> evenSchedule = getSsauSchedule(evenElement, true);
-        Map<DayOfWeek, ScheduleDay> oddSchedule = getSsauSchedule(oddElement, false);
-        Schedule schedule = new Schedule();
-        schedule.setEvenWeekSchedule(evenSchedule);
-        schedule.setOddWeekSchedule(oddSchedule);
-        schedule.setSnapshotDate(new Date());
-        log.info("Mapping SSAU data to schedule {} finished", schedule);
-        return schedule;
-    }
-
-    @Override
     public Schedule mapBmstuToSchedule(BmstuApiResponse bmstuApiResponse) {
         log.info("Start mapping BMSTU data to schedule");
         Schedule schedule = new Schedule();
@@ -122,65 +97,6 @@ public class MappingServiceImpl implements MappingService {
         schedule.setOddWeekSchedule(oddWeekSchedule);
         log.info("Mapping BMSTU data to schedule {} finished", schedule);
         return schedule;
-    }
-
-    @Override
-    public Schedule mapPgupsToSchedule(List<Document> documents) {
-        log.info("Start mapping PGUPS data to schedule");
-
-        Schedule schedule = new Schedule();
-        schedule.setSnapshotDate(new Date());
-        schedule.setEvenWeekSchedule(parsePgupsSchedule(documents.get(0), true));
-        schedule.setOddWeekSchedule(parsePgupsSchedule(documents.get(1), false));
-
-        log.info("Mapping PGUPS data to schedule {} finished", schedule);
-        return schedule;
-    }
-
-    private Map<DayOfWeek, ScheduleDay> parsePgupsSchedule(Document document, boolean isEvenWeek) {
-        Map<DayOfWeek, ScheduleDay> schedule = new LinkedHashMap<>();
-        Elements scheduleDays = document.getElementsByTag("tbody");
-
-        LocalDate baseWeekStart = LocalDate.now().with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
-        final LocalDate weekStart = isEvenWeek ? baseWeekStart : baseWeekStart.plusWeeks(1);
-
-        scheduleDays.forEach(day -> {
-            DayOfWeek dayOfWeek = parseDayOfWeek(day.getElementsByClass("kt-font-dark").text().toLowerCase());
-            LocalDate date = weekStart.with(dayOfWeek);
-            schedule.put(dayOfWeek, parsePgupsScheduleDay(day, date));
-        });
-
-        return schedule;
-    }
-
-    private ScheduleDay parsePgupsScheduleDay(Element dayElement, LocalDate date) {
-        Map<TimeSheet, List<ScheduleObject>> lessons = new LinkedHashMap<>();
-
-        dayElement.getElementsByTag("tr").forEach(lesson -> {
-            String[] timeRange = lesson.getElementsByClass("text-center kt-shape-font-color-4")
-                    .text()
-                    .split("—");
-            TimeSheet timeSheet = new TimeSheet(timeRange[0].trim(), timeRange[1].trim());
-            lessons.put(timeSheet, parsePgupsScheduleObjects(lesson));
-        });
-
-        ScheduleDay scheduleDay = new ScheduleDay();
-        scheduleDay.setLocalDate(date);
-        scheduleDay.setLessons(lessons);
-        return scheduleDay;
-    }
-
-    private List<ScheduleObject> parsePgupsScheduleObjects(Element lesson) {
-        List<ScheduleObject> scheduleObjects = new ArrayList<>();
-
-        ScheduleObject scheduleObject = new ScheduleObject();
-        scheduleObject.setName(lesson.getElementsByClass("mr-1").text().trim());
-        scheduleObject.setPlace(lesson.select("a[href^=https://rasp.pgups.ru/schedule/room]").text().trim());
-        scheduleObject.setType(ScheduleType.returnTypeByPgupsName(lesson.select("span[class^=badge]").text().trim()));
-        scheduleObject.setTeacher(lesson.select("div > a[href^=https://rasp.pgups.ru/schedule/teacher].kt-link").text().trim());
-
-        scheduleObjects.add(scheduleObject);
-        return scheduleObjects;
     }
 
     @Override
@@ -344,83 +260,6 @@ public class MappingServiceImpl implements MappingService {
         );
 
         return scheduleTypeMap.getOrDefault(bmstuType, ScheduleType.UNKNOWN);
-    }
-
-    private Map<DayOfWeek, ScheduleDay> getSsauSchedule(Element scheduleElement, boolean isEvenWeek) {
-        Map<DayOfWeek, ScheduleDay> scheduleDayMap = new LinkedHashMap<>();
-
-        LocalDate baseWeekStart = LocalDate.now().with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY));
-        final LocalDate startOfWeek = isEvenWeek ? baseWeekStart : baseWeekStart.plusWeeks(1);
-
-        for (DayOfWeek day : DayOfWeek.values()) {
-            if (day != DayOfWeek.SUNDAY) {
-                LocalDate date = startOfWeek.with(day);
-                scheduleDayMap.put(day, getSsauScheduleDay(day, scheduleElement, date));
-            }
-        }
-        return scheduleDayMap;
-    }
-
-    private ScheduleDay getSsauScheduleDay(DayOfWeek dayOfWeek, Element element, LocalDate date) {
-        Elements scheduleItemElements = element.getElementsByClass("schedule__item");
-        List<Element> ssauTimeSheets = element.getElementsByClass("schedule__time");
-        List<Element> ssauDayOfWeek = scheduleItemElements.subList(1, 7);
-        List<Element> ssauLessons = scheduleItemElements.subList(7, scheduleItemElements.size());
-
-        ScheduleDay scheduleDay = new ScheduleDay();
-        Map<TimeSheet, List<ScheduleObject>> timeSheetListMap = new LinkedHashMap<>();
-
-        for (int i = 0; i < ssauTimeSheets.size(); i++) {
-            String startTime = ssauTimeSheets.get(i).getElementsByClass("schedule__time-item").get(0).text();
-            String endTime = ssauTimeSheets.get(i).getElementsByClass("schedule__time-item").get(1).text();
-            TimeSheet timeSheet = new TimeSheet(startTime, endTime);
-
-            int currentElement = (i * ssauDayOfWeek.size()) + dayOfWeek.getValue() - 1;
-
-            List<ScheduleObject> scheduleObjects;
-            if (currentElement < ssauLessons.size()) {
-                scheduleObjects = getSsauScheduleObject(ssauLessons.get(currentElement));
-            } else {
-                scheduleObjects = new ArrayList<>();
-            }
-
-            timeSheetListMap.put(timeSheet, scheduleObjects);
-        }
-        scheduleDay.setLocalDate(date);
-        scheduleDay.setLessons(timeSheetListMap);
-
-        return scheduleDay;
-    }
-
-    private List<ScheduleObject> getSsauScheduleObject(Element element) {
-        List<ScheduleObject> scheduleObjects = new ArrayList<>();
-        List<Element> scheduleLessons = element.getElementsByClass("schedule__lesson");
-        for (Element scheduleLesson : scheduleLessons) {
-            ScheduleObject scheduleObject = new ScheduleObject();
-            scheduleObject.setType(ScheduleType.returnTypeByRuName(scheduleLesson.getElementsByClass("schedule__lesson-type-chip").text()));
-            scheduleObject.setName(scheduleLesson.getElementsByClass("schedule__discipline").text());
-            scheduleObject.setPlace(scheduleLesson.getElementsByClass("schedule__place").text());
-            scheduleObject.setTeacher(scheduleLesson.getElementsByClass("schedule__teacher").text());
-            scheduleObject.setGroups(getSsauScheduleGroups(scheduleLesson.getElementsByClass("schedule__groups")));
-            scheduleObjects.add(scheduleObject);
-        }
-        return scheduleObjects;
-    }
-
-    private List<String> getSsauScheduleGroups(Elements scheduleGroupsElement) {
-        List<String> groupNames = new ArrayList<>();
-        List<Element> aElements = scheduleGroupsElement.get(0).getElementsByTag("a");
-        List<Element> spanElements = scheduleGroupsElement.get(0).getElementsByTag("span");
-        if (aElements.isEmpty()) {
-            for (Element spanElement : spanElements) {
-                groupNames.add(spanElement.text());
-            }
-        } else {
-            for (Element aElement : aElements) {
-                groupNames.add(aElement.text());
-            }
-        }
-        return groupNames;
     }
 
     private Map<DayOfWeek, ScheduleDay> createSseuWeekScheduleWithoutLessons(SseuApiResponse sseuSchedule) {
